cache: Instrument hit/miss in prometheus
